Add Changes to EF Core And Validation

+++ Application/TodoItems/Commands/Delete.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Domain.Entities;
@@ -9,6 +10,7 @@ namespace Application.TodoItems.Commands
     public class DeleteTodoCommand : ICommand
     {
         public int TodoId { get; set; }
+        public string ListId { get; set; }
     }
 
     public class DeleteTodoItemHandler : ICommandHandler<DeleteTodoCommand>
@@ -22,9 +24,15 @@ namespace Application.TodoItems.Commands
 
         public async Task HandleAsync(DeleteTodoCommand command, CancellationToken token = default)
         {
+            if (!Guid.TryParse(command.ListId, out _))
+            {
+                throw new ArgumentException("ListId must be a Guid");
+            }
+            
             var entity = new TodoItem
             {
-                TodoId = command.TodoId
+                TodoId = command.TodoId,
+                ListId = command.ListId
             };
 
             await _todoItemRepository.DeleteAsync(entity, token);
